# api-gateway/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MessageBox API Gateway",
    "description": "API gateway to interact with the MessageBox backend micro-services set.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Development server",
      "variables": {}
    }
  ],
...
  "paths": {
...
    "/Users/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get user by id",
        "description": "Get user details by their user id.",
        "operationId": "GetUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User's unique ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched the user's details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "Update user by id",
        "description": "Update the user details by the their ID.",
        "operationId": "UpdateUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User's unique ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Updated user details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete user by id",
        "description": "Delete a user by their user id.<br /><br />The initial delete is a soft delete, which the user being set to inactive and the email being randomized.<br />A second subsequent request will perform a permanent deletion from the database.",
        "operationId": "DeleteUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User's unique ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessagePair"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/Users/{id}/ChangePassword": {
      "post": {
        "tags": ["User"],
        "summary": "Change user's password",
        "description": "Change a user's password.",
        "operationId": "ChangeUserPasswordById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User's unique ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "New password",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessagePair"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/Users/ConfirmUser": {
      "post": {
        "tags": ["User"],
        "summary": "Confirm the user account using the token sent via email.",
        "description": "Confirm the user account using the token sent via email.",
        "operationId": "ConfirmUserWithEmailToken",
        "parameters": [],
        "requestBody": {
          "description": "Body required to change the user's password",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully changed the user's password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessagePair"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/Users/ConfirmUser/ResendConfirmationEmail": {
      "post": {
        "tags": ["User"],
        "summary": "Resend the account confirmation email.",
        "description": "Resend the account confirmation email.",
        "operationId": "ResendConfirmationEmail",
        "parameters": [],
        "requestBody": {
          "description": "Body required to change the user's password",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "default": "msc@msc.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully resent the account confirmation email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessagePair"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/Users/ResetPassword/RequestEmail": {
      "post": {
        "tags": ["User"],
        "summary": "Request an email to reset the user password.",
        "description": "Request an email to reset the user password.",
        "operationId": "RequestPasswordResetWithEmail",
        "parameters": [],
        "requestBody": {
          "description": "Body required to change the user's password",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "default": "msc@msc.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully changed the user's password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessagePair"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/Users/ResetPassword/With2FA": {
      "post": {
        "tags": ["User"],
        "summary": "Change the user's password using the 2FA code.",
        "description": "Change the user's password using the 2FA code.",
        "operationId": "ChangePasswordWith2FA",
        "parameters": [],
        "requestBody": {
          "description": "Body required to change the user's password",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "default": "msc@msc.com"
                  },
                  "password": {
                    "type": "string",
                    "default": "123"
                  },
                  "code": {
                    "type": "string",
                    "default": "123456"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully changed the user's password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessagePair"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/Users/ResetPassword/WithToken": {
      "post": {
        "tags": ["User"],
        "summary": "Change the user's password using a token.",
        "description": "Change the user's password using a token.",
        "operationId": "ChangePasswordWithToken",
        "parameters": [],
        "requestBody": {
          "description": "Body required to change the user's password",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "password": {
                    "type": "string",
                    "default": "123"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully changed the user's password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessagePair"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/Users/Profile": {
      "get": {
        "tags": ["User"],
        "summary": "Get the logged-in user's details.",
        "description": "Get the logged-in user's details.",
        "operationId": "GetLoggedInUserProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged-in user profile details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "put": {
        "tags": ["User"],
        "summary": "Update user the logged-in user profile",
        "description": "Update the logged-in user.",
        "operationId": "UpdateLoggedInUserProfile",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Updated user details",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "UnauthorizedError": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Failed validation on the server",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/FieldValidationError"
            }
          }
        }
      }
    },
    "schemas": {
      "FieldValidationError": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "propertyPath": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "SuccessMessagePair": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "message": {
            "type": "string",
            "default": "Status of the request/operation"
          }
        }
      },
...
      "UserSchema": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "isActive": {
            "type": "boolean",
            "default": false
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": []
}